Changed TIER from an unsigned int to an unsigned long long, and fixed any warnings associated with the change.
git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanClassic/trunk@3258 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ gamesman/src/core/types.h
@@ -27,7 +27,7 @@ typedef char* 			STRING;
 typedef void* 			GENERIC_PTR;
 
 // Tier Gamesman typedefs
-typedef unsigned int TIER;
+typedef unsigned long long TIER;
 typedef POSITION TIERPOSITION; /* Used to differentiate "tier-specific" positions */
 typedef int UNDOMOVE; /* Just like MOVE but used specifically for undo-moves */
